requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
faster-whisper==1.1.0
yt-dlp==2024.12.23
ffmpeg-python==0.2.0
python-multipart==0.0.20
pydantic==2.10.6
aiofiles==24.1.0
requests==2.32.3
replicate==0.34.1
python-dotenv==1.0.1
